pyproject.toml
[project]
name = "photo-gen"
version = "0.2.0"
description = "AI-powered scene photo fusion"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.2.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.18.0",
    "opencv-python>=4.9.0",
    "Pillow>=10.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "numpy>=1.26.0",
    "volcengine-python-sdk[ark]>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graph", "services"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
]
